requirements.txt
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
mysql-connector-python==8.1.0
python-dotenv==1.0.0
schedule==1.2.0
pyyaml==6.0.1
requests==2.31.0
openpyxl==3.1.2
pytest==7.4.0
pytest-cov==4.1.0
faker==19.3.1
tqdm==4.66.1
colorlog==6.7.0